Remove auto-execution after edit of files
Commit 931750b introduced an update to the `\e` command to accept a
filename argument. On edit (ie: after closing the editor), the contents
would get executed in the current namespace, replicating the behavior of
editing the session history.

However, this is undesirable behavior since firstly, the cases in which
we edit files from the prompt is more often to 'read' the file rather
than to make changes and secondly even if we were to make changes, we
would like to execute only the changes in the current namespace (as
opposed to the entire file).

Also, we always have the option to call `exec` (/`execfile` in python2)
to execute the file.

* might provide for a new command to execute-lines-start-to-end in a
file, if we see the need to.