Fix long standing path completion annoyance
When completing pathnames within quotes, readline would autoclose the
quote on completion of directory names. This would happen because
the last completion was matched and readline (correctly) assumed quote
ought to be close.

The fix is so simple yet was elusive to almost everybody who faces it.
For example:

    https://github.com/ipython/ipython/issues/1172

We simply reset the completion matches to drill further down in the
directory if directory completion was ...ehe ..completed.