When executing lines from an edited buffer, stop when an exception occurs
This patchset fixes the annoying behavior that all lines of an edited buffer
would be executed. Although this was similar behavior, for instance, if we
were to copy/paste lines into the prompt, since in the case of the EDIT_CMD,
we are the ones calling exec, continuing execution after an exception is
avoidable.

The lines that are not executed after an exception are printed out nonetheless
in a different color.